chore: moved specs
