+++ src/libs/coin-validator.spec.ts
@@ -1,3 +1,11 @@
+/**
+ * `coin-validator.spec.ts`
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2021-01-22 optimized
+ *
+ * @copyright (C) 2021 LemonCloud Co Ltd. - All Rights Reserved.
+ */
 import { CoinValidator } from '../../src/';
 
 describe('CoinValidator', () => {
