+++ src/libs/date-helper.spec.ts
@@ -1,3 +1,11 @@
+/**
+ * `date-helper.spec.ts`
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2021-01-22 optimized
+ *
+ * @copyright (C) 2021 LemonCloud Co Ltd. - All Rights Reserved.
+ */
 import { DateHelper } from '../../src/';
 
 describe('DateHelper', () => {
